Graphql-server directory restructuring. (#2043)
* Restructuring graphql server
