+++ .circleci/config.yml
@@ -64,7 +64,7 @@ jobs:
       - run:
           name: Build graphql server docker image
           command: docker build . -f build/Dockerfile -t ${REPONAME}/${GRAPHQL_SERVER_IMAGE}:${IMGTAG}
-          working_directory: ~/project/litmus-portal/backend/graphql-server
+          working_directory: ~/project/litmus-portal/graphql-server
       - run:
           name: Save graphql server docker image
           command: docker save -o /tmp/workspace/${GRAPHQL_SERVER_IMAGE}.tar ${REPONAME}/${GRAPHQL_SERVER_IMAGE}:${IMGTAG}
