--- litmus-portal/backend/graphql-server/go.mod
@@ -1,24 +1,0 @@
-module github.com/litmuschaos/litmus/litmus-portal/backend/graphql-server
-
-go 1.14
-
-require (
-	github.com/99designs/gqlgen v0.11.3
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible
-	github.com/google/uuid v1.1.1
-	github.com/gorilla/mux v1.7.4
-	github.com/gorilla/websocket v1.2.0
-	github.com/imdario/mergo v0.3.11 // indirect
-	github.com/jinzhu/copier v0.0.0-20190924061706-b57f9002281a
-	github.com/pkg/errors v0.9.1
-	github.com/rs/cors v1.6.0
-	github.com/tidwall/sjson v1.1.1
-	github.com/vektah/gqlparser/v2 v2.0.1
-	go.mongodb.org/mongo-driver v1.3.5
-	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9 // indirect
-	golang.org/x/time v0.0.0-20200630173020-3af7569d3a1e // indirect
-	k8s.io/api v0.18.6
-	k8s.io/apimachinery v0.18.6
-	k8s.io/client-go v0.18.6
-	k8s.io/utils v0.0.0-20200603063816-c1c6865ac451 // indirect
-)
