+++ litmus-portal/graphql-server/gqlgen.yml
