+++ litmus-portal/graphql-server/utils/misc.go
